Minor updates
Basically a clean up to focus on what we want to use straight out of the gate and removing anything else. We can always add to this in the future. The only thing devs will need to do different is use the new `improvement/` branch prefix for improvements. So now there will be three different types of branches easily determined by the label given to the associated card in Trello:
- `feature/`
- `improvement/`
- `bugfix/`